[Update] BackPress 기능 추상화
EndPointFragment를 만들어 BackPress 기능을 추상화를 시킴. 불필요한 코드 줄음
