+++ presentation/src/main/java/com/mohaeyo/mohae/base/EndPointFragment.kt
@@ -1,0 +1,37 @@
+package com.mohaeyo.mohae.base
+
+import android.content.Context
+import androidx.activity.OnBackPressedCallback
+import androidx.databinding.ViewDataBinding
+import com.mohaeyo.mohae.backButtonSubject
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
+import org.jetbrains.anko.support.v4.toast
+
+abstract class EndPointFragment<V: ViewDataBinding>: DataBindingFragment<V>() {
+
+    private val backButtonSubjectDisposable: Disposable = backButtonSubject
+        .buffer(2, 1)
+        .observeOn(AndroidSchedulers.mainThread())
+        .subscribe {
+            if (it[1] - it[0] <= 1500) activity?.finish()
+            else toast("뒤로가기 버튼을 한 번 더 누르시면 종료됩니다.")
+        }
+
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+
+        requireActivity().onBackPressedDispatcher.addCallback(
+            this, object: OnBackPressedCallback(true) {
+                override fun handleOnBackPressed() {
+                    backButtonSubject.onNext(System.currentTimeMillis())
+                }
+            })
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        backButtonSubjectDisposable.dispose()
+    }
+
+}
