+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/feedback/FeedbackFragment.kt
@@ -7,40 +7,17 @@ import androidx.activity.OnBackPressedCallback
 import com.mohaeyo.mohae.R
 import com.mohaeyo.mohae.backButtonSubject
 import com.mohaeyo.mohae.base.DataBindingFragment
+import com.mohaeyo.mohae.base.EndPointFragment
 import com.mohaeyo.mohae.databinding.FragmentFeedbackBinding
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import org.jetbrains.anko.support.v4.toast
 
 
-class FeedbackFragment: DataBindingFragment<FragmentFeedbackBinding>() {
+class FeedbackFragment: EndPointFragment<FragmentFeedbackBinding>() {
 
     override val layoutId: Int
         get() = R.layout.fragment_feedback
-
-    private val backButtonSubjectDisposable: Disposable = backButtonSubject
-        .buffer(2, 1)
-        .observeOn(AndroidSchedulers.mainThread())
-        .subscribe {
-            if (it[1] - it[0] <= 1500) activity?.finish()
-            else toast("뒤로가기 버튼을 한 번 더 누르시면 종료됩니다.")
-        }
-
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-
-        requireActivity().onBackPressedDispatcher.addCallback(
-            this, object: OnBackPressedCallback(true) {
-                override fun handleOnBackPressed() {
-                    backButtonSubject.onNext(System.currentTimeMillis())
-                }
-            })
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        backButtonSubjectDisposable.dispose()
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
