+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/group/GroupFragment.kt
@@ -1,16 +1,23 @@
 package com.mohaeyo.mohae.ui.fragment.main.group
 
+import android.content.Context
 import android.os.Bundle
 import android.view.View
+import androidx.activity.OnBackPressedCallback
 import androidx.lifecycle.ViewModelProviders
 import com.mohaeyo.mohae.R
+import com.mohaeyo.mohae.backButtonSubject
 import com.mohaeyo.mohae.base.DataBindingFragment
+import com.mohaeyo.mohae.base.EndPointFragment
 import com.mohaeyo.mohae.databinding.FragmentGroupBinding
 import com.mohaeyo.mohae.viewmodel.main.group.GroupViewModel
 import com.mohaeyo.mohae.viewmodel.main.group.GroupViewModelFactory
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
+import org.jetbrains.anko.support.v4.toast
 import javax.inject.Inject
 
-class GroupFragment: DataBindingFragment<FragmentGroupBinding>() {
+class GroupFragment: EndPointFragment<FragmentGroupBinding>() {
 
     @Inject
     lateinit var factory: GroupViewModelFactory
