+++ presentation/src/main/java/com/mohaeyo/mohae/ui/fragment/main/mypage/MyPageFragment.kt
@@ -8,6 +8,7 @@ import androidx.fragment.app.Fragment
 import com.mohaeyo.mohae.R
 import com.mohaeyo.mohae.backButtonSubject
 import com.mohaeyo.mohae.base.DataBindingFragment
+import com.mohaeyo.mohae.base.EndPointFragment
 import com.mohaeyo.mohae.databinding.FragmentMypageBinding
 import dagger.android.AndroidInjector
 import dagger.android.DispatchingAndroidInjector
@@ -16,34 +17,10 @@ import io.reactivex.disposables.Disposable
 import org.jetbrains.anko.support.v4.toast
 import javax.inject.Inject
 
-class MyPageFragment: DataBindingFragment<FragmentMypageBinding>() {
+class MyPageFragment: EndPointFragment<FragmentMypageBinding>() {
 
     override val layoutId: Int
         get() = R.layout.fragment_mypage
-
-    private val backButtonSubjectDisposable: Disposable = backButtonSubject
-        .buffer(2, 1)
-        .observeOn(AndroidSchedulers.mainThread())
-        .subscribe {
-            if (it[1] - it[0] <= 1500) activity?.finish()
-            else toast("뒤로가기 버튼을 한 번 더 누르시면 종료됩니다.")
-        }
-
-    override fun onAttach(context: Context) {
-        super.onAttach(context)
-
-        requireActivity().onBackPressedDispatcher.addCallback(
-            this, object: OnBackPressedCallback(true) {
-                override fun handleOnBackPressed() {
-                    backButtonSubject.onNext(System.currentTimeMillis())
-                }
-            })
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        backButtonSubjectDisposable.dispose()
-    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
